+++ KaiFighter/KaiFighterGame/Objects/DynamicObjects/Characters/Enemies/Wizard.cs
@@ -16,7 +16,7 @@ namespace KaiFighterGame.Objects.DynamicObjects.Characters.Enemies
         private const string CollisionGroupString = "Wizard";
 
 
-        public Wizard(Vector2 position, string imageLocation, ObjectType objectType, Color? objColor, float scale, float rotation, float layerDepth, float movementSpeed, double damage, double health)
+        public Wizard(Vector2 position, string imageLocation, ObjectType objectType, Color? objColor, float scale, float rotation, float layerDepth, float movementSpeed, int damage, int health)
             : base(position, imageLocation, objectType, objColor, scale, rotation, layerDepth, movementSpeed, damage, health)
         {
