+++ KaiFighter/KaiFighterGame/Objects/DynamicObjects/Projectiles/Bullet.cs
@@ -8,9 +8,9 @@
     {
         private const string CollisionGroupString = "Bullet";
         private Vector2 targetDir;
-        private  bool isPlayerFire;
+        private bool isPlayerFire;
 
-        public Bullet(Vector2 position, string imageLocation, ObjectType objectType, Color? objColor, float scale, float rotation, float layerDepth, float movementSpeed, double damage, Vector2 targetDirection)
+        public Bullet(Vector2 position, string imageLocation, ObjectType objectType, Color? objColor, float scale, float rotation, float layerDepth, float movementSpeed, int damage, Vector2 targetDirection)
             : base(position, imageLocation, objectType, objColor, scale, rotation, layerDepth, movementSpeed)
         {
             this.Damage = damage;
@@ -18,7 +18,7 @@
            
         }
 
-        public double Damage { get; set; }
+        public int Damage { get; set; }
 
         // TODO - Change that !!!
         public bool IsPlayerFire
@@ -67,8 +67,6 @@
                     this.IsDestroyed = false;
                 }
             }
-           
-           
         }
     }
 }
